requirements.txt
ipython
jupyter
pytorch
torchvision
tqdm
numpy
matplotlib
